pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "uncloud"
version = "1.0.0"
description = "uncloud — CLI tool for ingesting and organizing Google Photos takeout exports and other photo archives"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [
    {name = "uncloud contributors"}
]
keywords = ["google-photos", "photo-organizer", "exif", "deduplication", "media"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Multimedia :: Graphics",
    "Topic :: Utilities",
]
dependencies = [
    "typer[all]>=0.12.3",
    "pydantic>=2.7.0",
    "PyYAML>=6.0.1",
    "Pillow>=10.2.0",
    "ImageHash>=4.3.1",
    "torchvision>=0.25.0",
    "torch>=2.0.0",
    "transformers>=4.30.0",
    "facenet-pytorch>=2.5.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.1.0",
]
build = [
    "pyinstaller>=6.0.0",
]

[project.scripts]
uncloud = "uncloud.cli:main"
uncloud-rebuild-index = "uncloud.cli:rebuild_index_main"

[project.urls]
Homepage = "https://github.com/MikeOfZen/uncloud"
Repository = "https://github.com/MikeOfZen/uncloud.git"
Issues = "https://github.com/MikeOfZen/uncloud/issues"
Documentation = "https://github.com/MikeOfZen/uncloud#readme"

[tool.setuptools.packages.find]
where = ["."]
include = ["uncloud*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_functions = "test_*"
addopts = "-v --tb=short"

[tool.coverage.run]
source = ["uncloud"]
branch = true

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "if __name__ == .__main__.:",
    "raise NotImplementedError",
]

[dependency-groups]
dev = [
    "pyinstaller>=6.18.0",
    "pytest>=9.0.2",
    "pytest-cov>=7.0.0",
    "ruff>=0.8.0",
    "pyright>=1.1.390",
    "codespell>=2.2.0",
]

# =============================================================================
# Ruff Configuration - Linting and Formatting
# =============================================================================

[tool.ruff]
target-version = "py310"
line-length = 100
src = ["uncloud", "tests"]

# Exclude patterns
exclude = [
    ".git",
    ".venv",
    "__pycache__",
    "build",
    "dist",
    "*.egg-info",
]

[tool.ruff.lint]
# Enable these rule sets
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # Pyflakes
    "I",      # isort (import sorting)
    "N",      # pep8-naming
    "UP",     # pyupgrade
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "SIM",    # flake8-simplify
    "ARG",    # flake8-unused-arguments
    "PTH",    # flake8-use-pathlib
    "ERA",    # eradicate (commented-out code)
    "RUF",    # Ruff-specific rules
    "PL",     # Pylint
    "PERF",   # Perflint (performance)
    "T20",    # flake8-print (catch print statements)
]

# Ignore specific rules
ignore = [
    "E501",    # Line too long - handled by formatter
    "PLR0913", # Too many arguments - sometimes needed
    "PLR2004", # Magic value in comparison - too noisy
    "SIM108",  # Use ternary operator - not always clearer
    "ARG001",  # Unused function argument - needed for protocols
    "ARG002",  # Unused method argument - needed for protocols
]

# Allow autofix for these
fixable = ["ALL"]
unfixable = []

[tool.ruff.lint.isort]
# Import sorting configuration
known-first-party = ["uncloud"]
combine-as-imports = true
force-wrap-aliases = true

[tool.ruff.lint.per-file-ignores]
# Test files can use assert and have unused variables
"tests/**/*.py" = ["S101", "ARG", "PLR2004", "T20"]
# Init files often have unused imports for re-export
"**/__init__.py" = ["F401"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"
docstring-code-format = true

# =============================================================================
# Pyright Configuration - Type Checking
# =============================================================================

[tool.pyright]
include = ["uncloud", "tests"]
exclude = [
    "**/__pycache__",
    "build",
    "dist",
    ".venv",
]

# Use standard mode for balance of strictness and practicality
typeCheckingMode = "standard"

# Python version
pythonVersion = "3.10"

# Report settings - catch common issues
reportMissingImports = true
reportMissingTypeStubs = false
reportIncompatibleVariableOverride = true
reportIncompatibleMethodOverride = true
reportUntypedFunctionDecorator = false
reportPrivateUsage = "warning"
reportUnusedImport = true
reportUnusedVariable = true
reportUnusedFunction = "warning"
reportDuplicateImport = true
reportOptionalMemberAccess = "warning"
reportOptionalSubscript = "warning"
reportOptionalCall = "warning"
reportOptionalIterable = "warning"
reportOptionalOperand = "warning"
reportTypedDictNotRequiredAccess = "warning"

# Strict settings we want even in standard mode
strictListInference = true
strictDictionaryInference = true
strictSetInference = true

# =============================================================================
# Codespell Configuration - Spelling
# =============================================================================

[tool.codespell]
skip = ".git,.venv,*.pyc,build,dist,*.egg-info,__pycache__"
ignore-words-list = "crate"

